clean up code organization a bit
