+++ crates/gitbutler-repo/src/traversal.rs
@@ -1,0 +1,97 @@
+use std::collections::HashSet;
+
+use anyhow::Result;
+use gix::{
+    revision::plumbing::{graph, merge_base::Flags},
+    revwalk::Graph,
+};
+
+/// Return commits on `from`'s first-parent chain, stopping before the first
+/// commit that is reachable from `stop_before`.
+///
+/// The returned commits are ordered from `from` backwards along the first-parent chain, excluding
+/// the first commit that is reachable from `stop_before` by ancestry.
+///
+/// This matches the semantics of a first-parent walk with `stop_before` hidden, but avoids the
+/// up-front hidden-side graph painting that makes `with_hidden(stop_before)` expensive in large
+/// repositories.
+pub fn first_parent_commit_ids_until(
+    repo: &gix::Repository,
+    from: gix::ObjectId,
+    stop_before: gix::ObjectId,
+) -> Result<Vec<gix::ObjectId>> {
+    let cache = repo.commit_graph_if_enabled()?;
+    let mut graph = repo.revision_graph(cache.as_ref());
+    first_parent_commit_ids_until_with_graph(repo, from, stop_before, &mut graph)
+}
+
+/// Like [`first_parent_commit_ids_until()`], but reuses `graph` across repeated ancestry queries.
+pub fn first_parent_commit_ids_until_with_graph(
+    repo: &gix::Repository,
+    from: gix::ObjectId,
+    stop_before: gix::ObjectId,
+    graph: &mut Graph<'_, '_, graph::Commit<Flags>>,
+) -> Result<Vec<gix::ObjectId>> {
+    let mut commit_ids = Vec::new();
+    let mut current = Some(from);
+
+    while let Some(commit_id) = current {
+        let reaches_hidden_history = match repo.merge_base_with_graph(commit_id, stop_before, graph)
+        {
+            Ok(merge_base) => merge_base.detach() == commit_id,
+            Err(gix::repository::merge_base_with_graph::Error::NotFound { .. }) => false,
+            Err(err) => return Err(err.into()),
+        };
+        if reaches_hidden_history {
+            break;
+        }
+
+        commit_ids.push(commit_id);
+        current = repo
+            .find_commit(commit_id)?
+            .parent_ids()
+            .next()
+            .map(|parent_id| parent_id.detach());
+    }
+
+    Ok(commit_ids)
+}
+
+/// Return commits reachable from `from` that are not reachable from `stop_before`.
+///
+/// This matches the semantics of walking `from` with `stop_before` hidden, but avoids
+/// the up-front hidden-side graph painting done by `with_hidden(stop_before)`.
+///
+/// Reuse `graph` across repeated ancestry queries for better performance.
+pub fn commit_ids_excluding_reachable_from_with_graph(
+    repo: &gix::Repository,
+    from: gix::ObjectId,
+    stop_before: gix::ObjectId,
+    graph: &mut Graph<'_, '_, graph::Commit<Flags>>,
+) -> Result<Vec<gix::ObjectId>> {
+    let mut commit_ids = Vec::new();
+    let mut seen = HashSet::new();
+    let mut to_visit = vec![from];
+
+    while let Some(commit_id) = to_visit.pop() {
+        if !seen.insert(commit_id) {
+            continue;
+        }
+
+        let reaches_hidden_history = match repo.merge_base_with_graph(commit_id, stop_before, graph)
+        {
+            Ok(merge_base) => merge_base.detach() == commit_id,
+            Err(gix::repository::merge_base_with_graph::Error::NotFound { .. }) => false,
+            Err(err) => return Err(err.into()),
+        };
+        if reaches_hidden_history {
+            continue;
+        }
+
+        commit_ids.push(commit_id);
+        let commit = repo.find_commit(commit_id)?;
+        to_visit.extend(commit.parent_ids().map(|parent_id| parent_id.detach()));
+    }
+
+    Ok(commit_ids)
+}
