+++ crates/gitbutler-repo/tests/fixtures/scenario/merge-history-prune.sh
