+++ crates/gitbutler-repo/tests/repo/main.rs
@@ -5,3 +5,4 @@ mod read_file_from_workspace_security;
 mod rebase;
 mod remotes;
 mod support;
+mod traversal;
